Ajout des libellés "d'aide" ou "de pret" sur les montant d'aide + déplacement d'AideDuree
